Create distracteddriver.py
The number of traffic accidents in Indonesia in the last three years tends to increase. The biggest factors causing accidents occur because of human errors such as drowsiness, chatting, driving while operating the telephone. The number of accident cases due to human negligence has encouraged the emergence of the use of technology to minimize accidents due to negligence of the driver. One way that can be done is to create an action recognition system that can automatically detect the condition of the driver. This study aims to build an action recognition model to detect driver negligence using convolutional neural network.
    The parameter values ​​were tested on CNN to get the optimal model. The best accuracy of the model is generated with RGB image input using the learning rate 0.002, dropout rate 0.1 parameter, and the convolution kernel with 32 and 32x32 sizes. The model is able to detect driver negligence based on the introduction of distracted driver action with the best average validation accuracy of 99 %.